Moved TopologyUtils to a new unittest case
