+++ CMakeLists.txt
@@ -479,7 +479,7 @@ install(TARGETS _GEOMAlgo DESTINATION ${PYTHONOCC_INSTALL_DIRECTORY}/Core )
 install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/src/Extend DESTINATION ${PYTHONOCC_INSTALL_DIRECTORY} )
 
 # install Utils
-install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/Wrapper/wrapper_utils.py DESTINATION ${PYTHONOCC_INSTALL_DIRECTORY})
+install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/src/Wrapper DESTINATION ${PYTHONOCC_INSTALL_DIRECTORY})
 
 # install deprecated modules
 file(GLOB DEPRECATED_MODULES ${CMAKE_CURRENT_SOURCE_DIR}/src/SWIG_files/deprecated_modules/*.py)
