+++ src/Extend/TopologyUtils.py
@@ -544,21 +544,6 @@ def discretize_edge(a_topods_edge, deflection=0.5):
         points.append(p.Coord())
     return points
 
-def test_discretize_edge():
-    from OCC.Core.BRepPrimAPI import BRepPrimAPI_MakeTorus
-    tor = BRepPrimAPI_MakeTorus(50, 20).Shape()
-    topo = TopologyExplorer(tor)
-    for edge in topo.edges():
-        discretize_edge(edge)
-
-
-def test_discretize_wire():
-    from OCC.Core.BRepPrimAPI import BRepPrimAPI_MakeTorus
-    tor = BRepPrimAPI_MakeTorus(50, 20).Shape()
-    topo = TopologyExplorer(tor)
-    for wire in topo.wires():
-        discretize_wire(wire)
-
 #
 # TopoDS_Shape type utils
 #
@@ -612,157 +597,3 @@ def is_compound(topods_shape):
 def is_compsolid(topods_shape):
     return topods_shape.ShapeType() == TopAbs_COMPSOLID
 
-
-#
-# Tests
-#
-
-def get_test_box_shape():
-    return BRepPrimAPI_MakeBox(10, 20, 30).Shape()
-
-
-def get_test_sphere_shape():
-    return BRepPrimAPI_MakeSphere(10.).Shape()
-
-def test_loop_faces():
-    i = 0
-    for face in topo.faces():
-        i += 1
-        assert(isinstance(face, TopoDS_Face))
-    assert i == 6
-
-
-def test_loop_edges():
-    i = 0
-    for face in topo.edges():
-        i += 1
-        assert(isinstance(face, TopoDS_Edge))
-    assert i == 12
-
-
-def number_of_topological_entities():
-    assert topo.number_of_faces() == 6
-    assert topo.number_of_edges() == 12
-    assert topo.number_of_vertices() == 8
-    assert topo.number_of_wires() == 6
-    assert topo.number_of_solids() == 1
-    assert topo.number_of_shells() == 1
-    assert topo.number_of_compounds() == 0
-    assert topo.number_of_comp_solids() == 0
-
-
-def test_nested_iteration():
-    '''check nested looping'''
-    for f in topo.faces():
-        for e in topo.edges():
-            assert isinstance(f, TopoDS_Face)
-            assert isinstance(e, TopoDS_Edge)
-
-
-def test_kept_reference():
-    '''did we keep a reference after looping several time through a list
-    of topological entities?'''
-    _tmp = []
-    _faces = [i for i in topo.faces()]
-    for f in _faces:
-        _tmp.append(f.IsNull() == 0)
-    for f in _faces:
-        _tmp.append(f.IsNull() == 0)
-    assert all(_tmp)
-
-
-def test_edge_face():
-    edg = next(topo.edges())
-    face = next(topo.faces())
-    faces_from_edge = [i for i in topo.faces_from_edge(edg)]
-    assert len(faces_from_edge) == topo.number_of_faces_from_edge(edg)
-    edges_from_face = [i for i in topo.edges_from_face(face)]
-    assert len(edges_from_face) == topo.number_of_edges_from_face(face)
-
-
-def test_edge_wire():
-    edg = next(topo.edges())
-    wire = next(topo.wires())
-    wires_from_edge = [i for i in topo.wires_from_edge(edg)]
-    assert len(wires_from_edge) == topo.number_of_wires_from_edge(edg)
-    edges_from_wire = [i for i in topo.edges_from_wire(wire)]
-    assert len(edges_from_wire) == topo.number_of_edges_from_wire(wire)
-
-
-def test_vertex_edge():
-    vert = next(topo.vertices())
-    edge = next(topo.edges())
-    verts_from_edge = [i for i in topo.vertices_from_edge(edge)]
-    assert len(verts_from_edge) == topo.number_of_vertices_from_edge(edge)
-    edges_from_vert = [i for i in topo.edges_from_vertex(vert)]
-    assert len(edges_from_vert) == topo.number_of_edges_from_vertex(vert)
-
-
-def test_vertex_face():
-    vert = next(topo.vertices())
-    face = next(topo.faces())
-    faces_from_vertex = [i for i in topo.faces_from_vertex(vert)]
-    assert len(faces_from_vertex) == topo.number_of_faces_from_vertex(vert)
-    verts_from_face = [i for i in topo.vertices_from_face(face)]
-    assert len(verts_from_face) == topo.number_of_vertices_from_face(face)
-
-
-def test_face_solid():
-    face = next(topo.faces())
-    solid = next(topo.solids())
-    faces_from_solid = [i for i in topo.faces_from_solids(solid)]
-    assert len(faces_from_solid) == topo.number_of_faces_from_solids(solid)
-    solids_from_face = [i for i in topo.solids_from_face(face)]
-    assert len(solids_from_face) == topo.number_of_solids_from_face(face)
-
-
-def test_wire_face():
-    wire = next(topo.wires())
-    face = next(topo.faces())
-    faces_from_wire = [i for i in topo.faces_from_wire(wire)]
-    assert len(faces_from_wire) == topo.number_of_faces_from_wires(wire)
-    wires_from_face = [i for i in topo.wires_from_face(face)]
-    assert len(wires_from_face) == topo.number_of_wires_from_face(face)
-
-
-def test_edges_out_of_scope():
-    # check pointers going out of scope
-    face = next(topo.faces())
-    _edges = []
-    for edg in TopologyExplorer(face).edges():
-        _edges.append(edg)
-    for edg in _edges:
-        assert not edg.IsNull()
-
-
-def test_wires_out_of_scope():
-    # check pointers going out of scope
-    wire = next(topo.wires())
-    _edges, _vertices = [], []
-    for edg in WireExplorer(wire).ordered_edges():
-        _edges.append(edg)
-    for edg in _edges:
-        assert not edg.IsNull()
-    for vert in WireExplorer(wire).ordered_vertices():
-        _vertices.append(vert)
-    for v in _vertices:
-        assert not v.IsNull()
-
-if __name__ == "__main__":
-    topo = TopologyExplorer(get_test_box_shape())
-    assert topo
-    test_kept_reference()
-    test_wire_face()
-    test_face_solid()
-    test_vertex_face()
-    test_vertex_edge()
-    test_edge_wire()
-    test_edge_face()
-    test_nested_iteration()
-    test_loop_edges()
-    test_loop_faces()
-    test_edges_out_of_scope()
-    test_wires_out_of_scope()
-    # tests discretize
-    test_discretize_edge()
-    test_discretize_wire()
